feat: recursive route loading ()
