+++ apps/web/src/pages/onboarding/(date-of-birth)/date-of-birth.ts
@@ -1,30 +1,34 @@
-import { createErrorSummary, formatZodErrors, getPreviousPage, getSessionDataForPage, processNameSubmission } from "@hmcts/onboarding";
+import { createErrorSummary, formatZodErrors, getPreviousPage, getSessionDataForPage, processDateOfBirthSubmission } from "@hmcts/onboarding";
 import type { Request, Response } from "express";
 import { ZodError } from "zod";
 
 const en = {
-  title: "What is your name?",
-  heading: "What is your name?",
-  firstNameLabel: "First name",
-  lastNameLabel: "Last name",
+  title: "What is your date of birth?",
+  heading: "What is your date of birth?",
+  hint: "For example, 27 3 1985",
+  dayLabel: "Day",
+  monthLabel: "Month",
+  yearLabel: "Year",
   back: "Back",
   continue: "Continue"
 };
 
 const cy = {
-  title: "Beth yw eich enw?",
-  heading: "Beth yw eich enw?",
-  firstNameLabel: "Enw cyntaf",
-  lastNameLabel: "Cyfenw",
+  title: "Beth yw eich dyddiad geni?",
+  heading: "Beth yw eich dyddiad geni?",
+  hint: "Er enghraifft, 27 3 1985",
+  dayLabel: "Diwrnod",
+  monthLabel: "Mis",
+  yearLabel: "Blwyddyn",
   back: "Yn ôl",
   continue: "Parhau"
 };
 
 export const GET = async (req: Request, res: Response) => {
-  const data = getSessionDataForPage(req.session, "name");
-  const backLink = getPreviousPage("name");
+  const data = getSessionDataForPage(req.session, "date-of-birth");
+  const backLink = getPreviousPage("date-of-birth");
 
-  res.render("onboarding/name", {
+  res.render("date-of-birth", {
     data,
     backLink,
     en,
@@ -34,22 +38,22 @@ export const GET = async (req: Request, res: Response) => {
 
 export const POST = async (req: Request, res: Response) => {
   try {
-    processNameSubmission(req.session, req.body);
+    processDateOfBirthSubmission(req.session, req.body);
 
     // Handle return parameter for change links
     const returnTo = req.query.return;
     if (returnTo === "summary") {
       res.redirect("/onboarding/summary");
     } else {
-      res.redirect("/onboarding/date-of-birth");
+      res.redirect("/onboarding/address");
     }
   } catch (error) {
     if (error instanceof ZodError) {
       const errors = formatZodErrors(error);
       const errorSummary = createErrorSummary(errors);
-      const backLink = getPreviousPage("name");
+      const backLink = getPreviousPage("date-of-birth");
 
-      res.render("onboarding/name", {
+      res.render("date-of-birth", {
         errors,
         errorSummary,
         data: req.body,
